refactor(frontend/context): move cart to cart folder
